backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
pymongo>=4.6
redis>=5.0
django-cors-headers>=4.3
gunicorn>=21.2
dnspython>=2.6
python-dotenv>=1.0
